# contracts/openapi.yaml
openapi: "3.0.0"
info:
  version: "0.1.0"
  title: ContextPilot API
  description: |
    Meeting intelligence REST API.
    Phase 0 placeholder: metadata and common schemas only.
    Endpoints added in BRD-02 and later domain specs.
servers:
  - url: /api/v1
    description: Base API path
...
security:
  - BearerAuth: []
  - CookieAuth: []
paths:
...
  /meetings/{id}/memory:
    get:
      operationId: getMeetingMemory
      summary: Get active memory for a meeting
      description: Returns the active memory version for a meeting, including processing state, briefing readiness, and full memory content. Gated by FF_ENABLE_MEETING_MEMORY_PROCESSING.
      tags:
        - memory
      security:
        - BearerAuth: []
        - CookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Active memory version
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Memory'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
        '403':
          description: Feature flag disabled or no memory available
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
        '404':
          description: Meeting not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
  /meetings/{id}/memory/versions:
    get:
      operationId: listMemoryVersions
      summary: List memory version history
      description: Returns all memory versions for a meeting, ordered by version number descending.
      tags:
        - memory
      security:
        - BearerAuth: []
        - CookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Version history
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MemoryVersionList'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
        '403':
          description: Feature flag disabled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
  /meetings/{id}/memory/versions/{versionNumber}:
    get:
      operationId: getMemoryVersion
      summary: Get a specific memory version
      description: Returns a specific memory version by version number.
      tags:
        - memory
      security:
        - BearerAuth: []
        - CookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: versionNumber
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Specific memory version
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Memory'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
        '404':
          description: Meeting or version not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
  /meetings/{id}/memory/reprocess:
    post:
      operationId: reprocessMemory
      summary: Queue memory reprocessing
      description: Queues a manual reprocess or includes/excludes prior memories as processing input. Gated by FF_ENABLE_MEETING_MEMORY_PROCESSING.
      tags:
        - memory
      security:
        - BearerAuth: []
        - CookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReprocessRequest'
      responses:
        '202':
          description: Reprocess job queued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReprocessAccepted'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
        '403':
          description: Feature flag disabled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
  /meetings/{id}/memory/state:
    get:
      operationId: getMemoryState
      summary: Get memory processing state
      description: Returns the current processing lifecycle state for a meeting's memory.
      tags:
        - memory
      security:
        - BearerAuth: []
        - CookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Processing state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MemoryProcessingState'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
  /meetings/{id}/memory/conflicts:
    get:
      operationId: listMemoryConflicts
      summary: List pending conflicts for review
      description: Returns all pending conflicts for a meeting's memory.
      tags:
        - memory
      security:
        - BearerAuth: []
        - CookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Conflict list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConflictList'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
        '403':
          description: Feature flag disabled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
  /meetings/{id}/memory/conflicts/{conflictId}/resolve:
    post:
      operationId: resolveMemoryConflict
      summary: Resolve a memory conflict
      description: Resolves a pending conflict with an evidence-backed resolution note. Creates a new memory version. Gated by FF_ENABLE_MEETING_MEMORY_PROCESSING.
      tags:
        - memory
      security:
        - BearerAuth: []
        - CookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: conflictId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConflictResolveRequest'
      responses:
        '200':
          description: Conflict resolved; new memory version returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Memory'
        '400':
          description: Invalid resolution note
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
        '403':
          description: Feature flag disabled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
        '404':
          description: Meeting or conflict not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
...
components:
...
  schemas:
    Problem:
      description: RFC 7807 error response
      type: object
      properties:
        type:
          type: string
          format: uri
          example: https://contextpilot.app/errors/problem
        title:
          type: string
          example: Resource Not Found
        status:
          type: integer
          example: 404
        detail:
          type: string
          example: Meeting with ID 12345 does not exist or you do not have access.
        instance:
          type: string
          format: uri
          example: /api/v1/meetings/12345
        request_id:
          type: string
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    CookieAuth:
      type: apiKey
      in: cookie
      name: session

  # ─── Memory Processing Schemas ───────────────────────────────────────────

    MemoryProcessingState:
      type: object
      description: Current processing lifecycle state for a meeting's memory.
      properties:
        meeting_id:
          type: string
          format: uuid
        processing_state:
          type: string
          enum: [not_processed, queued, processing, completed, completed_with_insufficient_evidence, failed, stale, reprocessing, retrying, retry_exhausted]
        active_version_number:
          type: integer
          nullable: true
          description: Present when a memory version exists.
        is_stale:
          type: boolean
          description: True when source meeting has been edited after the active version was created.
        latest_job:
          type: object
          nullable: true
          properties:
            job_id:
              type: string
              format: uuid
            status:
              type: string
              enum: [queued, processing, completed, completed_with_insufficient_evidence, failed, retrying, retry_exhausted]
            retry_count:
              type: integer
...
  securitySchemes: